+++ R/lencode_glm.R
@@ -222,17 +222,25 @@ print.step_lencode_glm <-
 #' @export
 tidy.step_lencode_glm <- function(x, ...) {
   if (is_trained(x)) {
+    if (length(x$mapping) == 0) {
+      res <- tibble(
+        terms = character(),
+        level = character(),
+        value = double()
+      )
+    } else {
     for (i in seq_along(x$mapping)) {
       x$mapping[[i]]$terms <- names(x$mapping)[i]
     }
     res <- bind_rows(x$mapping)
     names(res) <- gsub("^\\.\\.", "", names(res))
+    }
   } else {
     term_names <- sel2char(x$terms)
     res <- tibble(
+      terms = term_names,
       level = rep(na_chr, length(term_names)),
-      value = rep(na_dbl, length(term_names)),
-      terms = term_names
+      value = rep(na_dbl, length(term_names))
     )
   }
   res$id <- x$id