+++ R/pca_truncated.R
@@ -150,6 +150,7 @@ prep.step_pca_truncated <- function(x, training, info = NULL, ...) {
     }
   } else {
     prc_obj <- NULL
+    prc_obj$rotation <- matrix(nrow = 0, ncol = 0)
   }
 
   rownames(prc_obj$rotation) <- col_names
@@ -274,12 +275,20 @@ tidy.step_pca_truncated <- function(x, type = "coef", ...) {
       component = na_chr
     )
   } else {
-    type <- match.arg(type, c("coef", "variance"))
-    if (type == "coef") {
-      x$res <- x$res$rotation
-      res <- pca_coefs(x)
+    if (length(x$terms) == 0) {
+      res <- tibble(
+        terms = character(),
+        value = double(),
+        component = character()
+      )
     } else {
-      res <- pca_variances(x)
+      type <- match.arg(type, c("coef", "variance"))
+      if (type == "coef") {
+        x$res <- x$res$rotation
+        res <- pca_coefs(x)
+      } else {
+        res <- pca_variances(x)
+      }
     }
   }
   res$id <- x$id