+++ tests/testthat/test-discretize_cart.R
@@ -137,12 +137,12 @@ test_that("tidy method", {
 
   res <- tidy(cart_rec, number = 1)
   expect_equal(
-    res$variable,
+    res$terms,
     "all_predictors()"
   )
   expect_equal(
-    res$values,
-    NA_character_
+    res$value,
+    NA_real_
   )
 
   expect_snapshot({
@@ -155,7 +155,7 @@ test_that("tidy method", {
     rep("x", 2)
   )
   expect_equal(
-    res$values,
+    res$value,
     best_split
   )
 })
@@ -242,6 +242,43 @@ test_that("bake method errors when needed non-standard role columns are missing"
     bake(rec_trained, new_data = sim_tr_cls[, -1]),
     class = "new_data_missing_column"
   )
+})
+
+test_that("empty printing", {
+  rec <- recipe(mpg ~ ., mtcars)
+  rec <- step_discretize_cart(rec, outcome = "mpg")
+  
+  expect_snapshot(rec)
+  
+  rec <- prep(rec, mtcars)
+  
+  expect_snapshot(rec)
+})
+
+test_that("empty selection prep/bake is a no-op", {
+  rec1 <- recipe(mpg ~ ., mtcars)
+  rec2 <- step_discretize_cart(rec1, outcome = "mpg")
+  
+  rec1 <- prep(rec1, mtcars)
+  rec2 <- prep(rec2, mtcars)
+  
+  baked1 <- bake(rec1, mtcars)
+  baked2 <- bake(rec2, mtcars)
+  
+  expect_identical(baked1, baked2)
+})
+
+test_that("empty selection tidy method works", {
+  rec <- recipe(mpg ~ ., mtcars)
+  rec <- step_discretize_cart(rec, outcome = "mpg")
+  
+  expect <- tibble(terms = character(), value = double(), id = character())
+  
+  expect_identical(tidy(rec, number = 1), expect)
+  
+  rec <- prep(rec, mtcars)
+  
+  expect_identical(tidy(rec, number = 1), expect)
 })
 
 test_that("printing", {