+++ tests/testthat/test-embed.R
@@ -381,6 +381,51 @@ test_that("bake method errors when needed non-standard role columns are missing"
   )
 })
 
+test_that("empty printing", {
+  skip_on_cran()
+  skip_if(!is_tf_available())
+  
+  rec <- recipe(mpg ~ ., mtcars)
+  rec <- step_embed(rec, outcome = vars(mpg))
+  
+  expect_snapshot(rec)
+  
+  rec <- prep(rec, mtcars)
+  
+  expect_snapshot(rec)
+})
+
+test_that("empty selection prep/bake is a no-op", {
+  rec1 <- recipe(mpg ~ ., mtcars)
+  rec2 <- step_embed(rec1, outcome = vars(mpg))
+  
+  rec1 <- prep(rec1, mtcars)
+  rec2 <- prep(rec2, mtcars)
+  
+  baked1 <- bake(rec1, mtcars)
+  baked2 <- bake(rec2, mtcars)
+  
+  expect_identical(baked1, baked2)
+})
+
+test_that("empty selection tidy method works", {
+  rec <- recipe(mpg ~ ., mtcars)
+  rec <- step_embed(rec, outcome = vars(mpg))
+  
+  expect <- tibble(
+    terms = character(),
+    level = character(),
+    value = double(),
+    id = character()
+  )
+  
+  expect_identical(tidy(rec, number = 1), expect)
+  
+  rec <- prep(rec, mtcars)
+  
+  expect_identical(tidy(rec, number = 1), expect)
+})
+
 test_that("printing", {
   skip_on_cran()
   skip_if(!is_tf_available())