+++ tests/testthat/test-woe.R
@@ -282,17 +282,66 @@ test_that("tunable is setup to works with extract_parameter_set_dials works", {
 # Infrastructure ---------------------------------------------------------------
 
 test_that("bake method errors when needed non-standard role columns are missing", {
-  rec <- recipe(Status ~ ., data = credit_tr) %>%
-    step_discretize(Price) %>%
-    update_role(Price, new_role = "potato") %>%
+  rec <- recipe(credit_tr) %>%
+    step_woe(Job, Home, outcome = vars(Status)) %>%
+    update_role(Job, new_role = "potato") %>%
     update_role_requirements(role = "potato", bake = FALSE)
   
-  rec_trained <- prep(rec, training = credit_tr, verbose = FALSE)
+  suppressWarnings(
+    rec_trained <- prep(rec, training = credit_tr, verbose = FALSE)
+  )
   
   expect_error(
-    bake(rec_trained, new_data = credit_tr[, -14]),
+    bake(rec_trained, new_data = credit_tr[, -8]),
     class = "new_data_missing_column"
   )
+})
+
+test_that("empty printing", {
+  rec <- recipe(mpg ~ ., mtcars)
+  rec <- step_woe(rec, outcome = vars(mpg))
+  
+  expect_snapshot(rec)
+  
+  rec <- prep(rec, mtcars)
+  
+  expect_snapshot(rec)
+})
+
+test_that("empty selection prep/bake is a no-op", {
+  rec1 <- recipe(mpg ~ ., mtcars)
+  rec2 <- step_woe(rec1, outcome = vars(mpg))
+  
+  rec1 <- prep(rec1, mtcars)
+  rec2 <- prep(rec2, mtcars)
+  
+  baked1 <- bake(rec1, mtcars)
+  baked2 <- bake(rec2, mtcars)
+  
+  expect_identical(baked1, baked2)
+})
+
+test_that("empty selection tidy method works", {
+  rec <- recipe(mpg ~ ., mtcars)
+  rec <- step_woe(rec, outcome = vars(mpg))
+  
+  expect <- res <- tibble(
+    terms = character(),
+    value = character(),
+    n_tot = integer(),
+    n_bad = integer(),
+    n_good = integer(),
+    p_bad = double(),
+    p_good = double(),
+    woe = double(),
+    id = character()
+  )
+  
+  expect_identical(tidy(rec, number = 1), expect)
+  
+  rec <- prep(rec, mtcars)
+  
+  expect_identical(tidy(rec, number = 1), expect)
 })
 
 test_that("printing", {